requirements.txt
SQLAlchemy==1.4.25
Flask==2.0.2
flask_expects_json==1.6.0
Flask_SQLAlchemy==2.5.1
Werkzeug==2.0.2
jsonschema==3.2.0
pytest==6.2.5
PyJWT==2.2.0
